refactor: recipes module folder
