+++ packages/shared/src/recipes/ingredient.types.ts
@@ -1,0 +1,4 @@
+import type { z } from "zod";
+import type { ingredientSchema } from "./ingredient.schema.js";
+
+export type Ingredient = z.infer<typeof ingredientSchema>;
